modifs

+++ coevo/evolution_strategies.py
@@ -133,7 +133,7 @@ class ES:
     # Specific to algo 
     def populate(self):
         for i in range(self.n_pop):
-            new_genes = self.rng.standard_normal(self.d)
+            new_genes = self.rng.standard_normal(self.d, dtype=np.float32)
             self.population[i].genes = new_genes 
         return self
